made changes to routes and middlewares

+++ Modules/Products/deleteProduct.js
@@ -7,6 +7,9 @@ const deleteProduct = async (req, res) => {
 
   if (!validator.isMongoId(productId)) throw "Invalid Product ID";
 
+  const getProduct = await productsModel.findById(productId);
+  if (!getProduct) throw "Product does not exist";
+
   const deletedItem = await productsModel.findByIdAndDelete(productId);
 
   res.status(200).json({
